WooCommerce: Make api-plan general purpose
This takes the api-plan code that was originally developed under
state/ui/product and places it under state/api-plan. It also seeks to
make more general use of the code and adds a README to explain its
operation.

This is really just the reducer/action code, there are still selectors
to be written and more helper functions to be ported/written.