Eliminate race condition in form state load
The explicit `setState` call at the end of `componentWillMount` is unnecessary (because `loadFunction` generates an `OnNewState` event which sets the form state).

The `setState` call actually causes a race condition that had been previously masked by a slow (relative to local JS execution) API call hooked to `loadFunction`. The `loadFunction` naturally tended to return after the explicit `setState` call previously but when we switched to pulling from Redux state the behavior inverted which meant that the explicit `setState` actually destroyed the state we desired!